Split out date / time tests
